backend-python/requirements.txt
flask
flask-cors
rembg
onnxruntime
pillow
numpy
gunicorn app:app